+++ src/modelmanager.hpp
@@ -55,10 +55,8 @@ protected:
      */
     ModelManager(const std::string& modelCacheDirectory = "");
 
-public:
-    void details();
+    void logPluginConfiguration();
 
-protected:
     Status checkStatefulFlagChange(const std::string& modelName, bool configStatefulFlag);
 
     std::shared_ptr<ovms::Model> getModelIfExistCreateElse(const std::string& name, const bool isStateful);
